feat: Add Images API Route
Issue [#2](https://github.com/Bluette1/still-space/issues/2)
