+++ Gemfile
@@ -38,6 +38,7 @@ gem 'bootsnap', require: false
 group :development, :test do
   # See https://guides.rubyonrails.org/debugging_rails_applications.html#debugging-with-the-debug-gem
   gem 'debug', platforms: %i[mri windows]
+  gem 'rspec-rails', '~> 5.0.0'
   gem 'rubocop', require: false
 end
 
@@ -50,3 +51,4 @@ end
 
 gem 'devise', '~> 4.9'
 gem 'devise-api', '~> 0.1.3'
+gem 'rack-cors', require: 'rack/cors'
