+++ config/application.rb
@@ -28,5 +28,12 @@ module StillspaceApi
     # Middleware like session, flash, cookies can be added back manually.
     # Skip views, helpers and assets when generating a new resource.
     config.api_only = true
+
+    config.middleware.insert_before 0, Rack::Cors do
+      allow do
+        origins '*' # Allow requests from all origins, or specify your app's URL
+        resource '*', headers: :any, methods: %i[get post options]
+      end
+    end
   end
 end
